+++ spec/integration/resource/api500/synergy/server_hardware/create_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API300::Synergy::ServerHardware
+klass = OneviewSDK::API500::Synergy::ServerHardware
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  include_context 'integration api300 context'
+  include_context 'integration api500 context'
 
   describe '#add' do
     it 'is a pending test due to no synergy schematics containing iLOs'